[handleUserDidAssociation] move to hooks dir and fix user patch
